move const.go to constants package so it's available to internals
