+++ ApolloBuilder.go
@@ -259,12 +259,12 @@ func (b *Apollo) buildFakeWitnessSet() TransactionWitnessSet.TransactionWitnessS
 	plutusdata = append(plutusdata, b.datums...)
 	fakeVkWitnesses := make([]VerificationKeyWitness.VerificationKeyWitness, 0)
 	fakeVkWitnesses = append(fakeVkWitnesses, VerificationKeyWitness.VerificationKeyWitness{
-		Vkey:      FAKE_VKEY,
-		Signature: FAKE_SIGNATURE})
+		Vkey:      constants.FAKE_VKEY,
+		Signature: constants.FAKE_SIGNATURE})
 	for range b.requiredSigners {
 		fakeVkWitnesses = append(fakeVkWitnesses, VerificationKeyWitness.VerificationKeyWitness{
-			Vkey:      FAKE_VKEY,
-			Signature: FAKE_SIGNATURE})
+			Vkey:      constants.FAKE_VKEY,
+			Signature: constants.FAKE_SIGNATURE})
 	}
 	return TransactionWitnessSet.TransactionWitnessSet{
 		NativeScripts:  b.nativescripts,
@@ -832,7 +832,7 @@ func (a *Apollo) SetWalletFromMnemonic(mnemonic string) *Apollo {
 }
 
 // For use with key pairs generated by cardano-cli
-func (a *Apollo) SetWalletFromKeypair(vkey string, skey string, network Network) *Apollo {
+func (a *Apollo) SetWalletFromKeypair(vkey string, skey string, network constants.Network) *Apollo {
 	verificationKey_bytes, err := hex.DecodeString(vkey)
 	if err != nil {
 		fmt.Println("SetWalletFromKeypair: Failed to decode vkey")
@@ -850,7 +850,7 @@ func (a *Apollo) SetWalletFromKeypair(vkey string, skey string, network Network)
 	vkh, _ := verificationKey.Hash()
 
 	addr := Address.Address{}
-	if network == MAINNET {
+	if network == constants.MAINNET {
 		addr = Address.Address{StakingPart: nil, PaymentPart: vkh[:], Network: 1, AddressType: Address.KEY_NONE, HeaderByte: 0b01100001, Hrp: "addr"}
 	} else {
 		addr = Address.Address{StakingPart: nil, PaymentPart: vkh[:], Network: 0, AddressType: Address.KEY_NONE, HeaderByte: 0b01100000, Hrp: "addr_test"}
