+++ backends.go
@@ -1,6 +1,7 @@
 package apollo
 
 import (
+	"github.com/SundaeSwap-finance/apollo/constants"
 	"github.com/SundaeSwap-finance/apollo/txBuilding/Backend/BlockFrostChainContext"
 	"github.com/SundaeSwap-finance/apollo/txBuilding/Backend/FixedChainContext"
 )
@@ -11,32 +12,32 @@ func NewEmptyBackend() FixedChainContext.FixedChainContext {
 
 func NewBlockfrostBackend(
 	projectId string,
-	network Network,
+	network constants.Network,
 
 ) BlockFrostChainContext.BlockFrostChainContext {
 	switch network {
-	case MAINNET:
+	case constants.MAINNET:
 		return BlockFrostChainContext.NewBlockfrostChainContext(
-			BLOCKFROST_BASE_URL_MAINNET,
-			int(MAINNET),
+			constants.BLOCKFROST_BASE_URL_MAINNET,
+			int(constants.MAINNET),
 			projectId,
 		)
-	case TESTNET:
+	case constants.TESTNET:
 		return BlockFrostChainContext.NewBlockfrostChainContext(
-			BLOCKFROST_BASE_URL_TESTNET,
-			int(TESTNET),
+			constants.BLOCKFROST_BASE_URL_TESTNET,
+			int(constants.TESTNET),
 			projectId,
 		)
-	case PREVIEW:
+	case constants.PREVIEW:
 		return BlockFrostChainContext.NewBlockfrostChainContext(
-			BLOCKFROST_BASE_URL_PREVIEW,
-			int(TESTNET),
+			constants.BLOCKFROST_BASE_URL_PREVIEW,
+			int(constants.TESTNET),
 			projectId,
 		)
-	case PREPROD:
+	case constants.PREPROD:
 		return BlockFrostChainContext.NewBlockfrostChainContext(
-			BLOCKFROST_BASE_URL_PREPROD,
-			int(TESTNET),
+			constants.BLOCKFROST_BASE_URL_PREPROD,
+			int(constants.TESTNET),
 			projectId,
 		)
 	default:
